Fix Sourcefile to only update configuration if it does not already exist. Update bootstrapping script to make use paklib target 'configure' for the entire configuration process. Update utility 'conf' to not write any data if a given key is already set to the specified variable, saving more disk writes. Update addpayload and Tupfile to allow 'lite' builds, creating a simple bash script with a payload without any binary content at all, depending on lua and git and build systems on the target system.

+++ paklib/utils/conf.lua
@@ -14,6 +14,7 @@ conf.update = function(key,var,confile)
       t[k] = v
     end
   end
+  if t[key] == var then return true end -- avoid writing a variable already set
   t[key] = var
   f = io.open(confile,'w+b')
   f:flush()
